requirements.txt
cx_Oracle==8.3
pymongo==4.7.1
redis==5.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1